requirements.txt
anyio==3.6.1
asgiref==3.5.2
certifi==2022.5.18.1
click==8.1.3
cloudinary==1.29.0
dnspython==2.2.1
email-validator==1.2.1
fastapi==0.78.0
gunicorn==20.1.0
h11==0.13.0
idna==3.3
psycopg2-binary==2.9.3
pydantic==1.9.1
pydentic==0.0.1.dev3
PyJWT==2.4.0
python-decouple==3.6
python-dotenv==0.20.0
python-multipart==0.0.5
python-stdnum==1.17
six==1.16.0
sniffio==1.2.0
starlette==0.19.1
typing_extensions==4.2.0
urllib3==1.26.9
uvicorn==0.17.6